:zap: refactor(cdk): consolidate assistant api lambda functions to reduce cloudformation resources
Consolidated 7 Lambda functions (createAssistant, listAssistants, getAssistant, updateAssistant, deleteAssistant, createAssistantMessage, listAssistantMessages) into 2 unified handlers:
- assistantHandler.ts: handles all CRUD operations (create, list, get, update, delete)
- assistantMessageHandler.ts: handles message operations (create message, list messages)

This reduces CloudFormation stack resources from 513 to 498 (under the 500 limit).

All API endpoints remain unchanged - no breaking changes.
All functionality is identical to previous implementation.
