--- packages/cdk/lambda/listAssistants.ts
@@ -1,34 +1,0 @@
-import { APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda';
-import { listAssistants } from './repository/assistant';
-
-export const handler = async (
-  event: APIGatewayProxyEvent
-): Promise<APIGatewayProxyResult> => {
-  try {
-    const userId: string =
-      event.requestContext.authorizer!.claims['cognito:username'];
-
-    const exclusiveStartKey = event.queryStringParameters?.exclusiveStartKey;
-
-    const result = await listAssistants(userId, event, exclusiveStartKey);
-
-    return {
-      statusCode: 200,
-      headers: {
-        'Content-Type': 'application/json',
-        'Access-Control-Allow-Origin': '*',
-      },
-      body: JSON.stringify(result),
-    };
-  } catch (error) {
-    console.error('Error listing assistants:', error);
-    return {
-      statusCode: 500,
-      headers: {
-        'Content-Type': 'application/json',
-        'Access-Control-Allow-Origin': '*',
-      },
-      body: JSON.stringify({ message: 'Internal Server Error' }),
-    };
-  }
-};
